# Revert use of SkipPastAny in SkipWhiteSpace and optimize BufferReader.IsNext